requirements.txt
aiogram>=3.7        # Telegram-бот
loguru>=0.7
notifiers>=1.3
python-dotenv>=1.0
requests>=2.32
seleniumbase>=4.25
apscheduler==3.10.4
bs4>=0.0.1            # Beautiful Soup для парсинга HTML
beautifulsoup4>=4.12.0 # Актуальная версия Beautiful Soup
fake-useragent>=1.2.1  # Для генерации случайных User-Agent